fix: remove orig files in favor of v2 versions
